+++ NEWS.md
@@ -5,7 +5,7 @@
 - Renamed `se_mean_cat()` to `se_prop()`.
 - Renamed `se_mean_num()` to `se_mean()`.
 - Added `se_total_prop()` to estimate totals and proportions of categorical variables.
-- Added `se_total_comb()` and `se_mean_comb()` to estimate totals and means, respectively, for all possible grouping variable combinations.
+- Added `se_total_ogd()` and `se_mean_ogd()` to estimate totals and means, respectively, for all possible grouping variable combinations.
 
 # chensus 1.0.0